requirements.txt
langchain-openai
python-dotenv
pydantic
streamlit